requirements.txt
# requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
python-dotenv==1.0.1
openai==1.14.3
requests==2.31.0
moviepy==1.0.3
ffmpeg-python==0.2.0
pydantic==2.5.3
